Addition of report generation class to quiz package.
I have linked all through from login to viewchallenges, attemptchallenges and generation of the report.
